Time: 174 ms (98.34%) | Memory: 0B (100.00%) - LeetSync

+++ 1452-restaurant-growth/restaurant-growth.sql
@@ -4,15 +4,19 @@ WITH cte AS (
         visited_on,
         SUM(amount) AS amount
     FROM Customer
-    GROUP BY visited_on 
+    GROUP BY visited_on
     ORDER BY visited_on
 )
 
 SELECT
     visited_on,
-    SUM(amount) OVER(ROWS BETWEEN 6 PRECEDING AND CURRENT ROW) AS amount,
+    SUM(amount) OVER(
+        ROWS BETWEEN 6 PRECEDING AND CURRENT ROW
+    ) AS amount,
     ROUND(
-        AVG(amount) OVER(ROWS BETWEEN 6 PRECEDING AND CURRENT ROW)
+        AVG(amount::DECIMAL) OVER(
+            ROWS BETWEEN 6 PRECEDING AND CURRENT ROW
+        )
         ,2
     ) AS average_amount
 FROM cte
